requirements.txt
Flask>=2.3.0
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0
Werkzeug>=2.3.0
reportlab>=4.0.0
